WiP with relative paths.
