+++ agents/agent.py
@@ -19,19 +19,19 @@ from langgraph.prebuilt import create_react_agent
 from pydantic import ValidationError
 from trustcall import create_extractor
 
-from agents.agent_responses import AnalysisInsights, FilePath
-from agents.prompts import FILE_CLASSIFICATION_MESSAGE
 from agents.tools import CodeReferenceReader, CodeStructureTool, PackageRelationsTool, FileStructureTool, GetCFGTool, \
     MethodInvocationsTool, ReadFileTool
 from agents.tools.external_deps import ExternalDepsTool
 from agents.tools.read_docs import ReadDocsTool
 from static_analyzer.analysis_result import StaticAnalysisResults
+from static_analyzer.reference_resolve_mixin import ReferenceResolverMixin
 
 logger = logging.getLogger(__name__)
 
 
-class CodeBoardingAgent:
+class CodeBoardingAgent(ReferenceResolverMixin):
     def __init__(self, repo_dir: Path, static_analysis: StaticAnalysisResults, system_message: str):
+        super().__init__(repo_dir, static_analysis)
         self._setup_env_vars()
         self.llm = self._initialize_llm()
         self.extractor_llm = self._initialize_llm()
@@ -168,85 +168,6 @@ class CodeBoardingAgent:
             logger.error(f"Resource exhausted or parsing error, retrying... in 60 seconds: Type({type(e)}) {e}")
             time.sleep(60)
             return self._parse_response(prompt, response, return_type, max_retries - 1)
-
-    def fix_source_code_reference_lines(self, analysis: AnalysisInsights):
-        logger.info(f"Fixing source code reference lines for the analysis: {analysis.llm_str()}")
-        for component in analysis.components:
-            for reference in component.referenced_source_code:
-                # Check if the file is already resolved
-                if reference.reference_file is not None and os.path.exists(reference.reference_file):
-                    continue
-                matched = False
-                for lang in self.static_analysis.get_languages():
-                    try:
-                        qname = reference.qualified_name.replace("/", ".")
-                        node = self.static_analysis.get_reference(lang, qname)
-                        reference.reference_file = node.file_path
-                        reference.reference_start_line = node.line_start + 1  # match 1 based indexing
-                        reference.reference_end_line = node.line_end + 1  # match 1 based indexing
-                        reference.qualified_name = qname
-                        logger.info(
-                            f"[Reference Resolution] Matched {reference.qualified_name} in {lang} at {reference.reference_file}")
-                        matched = True
-                        break
-                    except (ValueError, FileExistsError) as e:
-                        # Try resolving with loose matching:
-                        qname = reference.qualified_name.replace("/", ".")
-                        _, node = self.static_analysis.get_loose_reference(lang, qname)
-                        if node is not None:
-                            reference.reference_file = node.file_path
-                            reference.reference_start_line = node.line_start + 1
-                            reference.reference_end_line = node.line_end + 1
-                            reference.qualified_name = qname
-                            logger.info(
-                                f"[Reference Resolution] Loosely matched {reference.qualified_name} in {lang} at {reference.reference_file}")
-                            matched = True
-                            break
-                        # before we give up let's retry with the file:
-                        logger.warning(
-                            f"[Reference Resolution] Reference {reference.qualified_name} not found in {lang}: {e}")
-                        if (reference.reference_file is not None) and (not reference.reference_file.startswith("/")):
-                            joined_path = os.path.join(self.repo_dir, reference.reference_file)
-                            if os.path.exists(joined_path):
-                                reference.reference_file = joined_path
-                                logger.info(
-                                    f"[Reference Resolution] File path matched for {reference.qualified_name} in {lang} at {reference.reference_file}")
-                                matched = True
-                                break
-                            else:
-                                reference.reference_file = None
-                        # Check if the code reference is a file path:
-                        file_path = reference.qualified_name.replace(".", "/")  # Get file path
-                        full_path = os.path.join(self.repo_dir, file_path)
-                        # This is the case when the reference is a file path but wrong:
-                        file_ref = ".".join(full_path.rsplit("/", 1))
-                        paths = [full_path, f"{file_path}.py", f"{file_path}.ts", f"{file_path}.tsx", file_ref]
-                        for path in paths:
-                            if os.path.exists(path):
-                                reference.reference_file = str(path)
-                                logger.info(
-                                    f"[Reference Resolution] Path matched for {reference.qualified_name} in {lang} at {reference.reference_file}")
-                                matched = True
-                                break
-                        if matched:
-                            break
-                        if (reference.reference_file is None) or (not matched):
-                            prompt = PromptTemplate(template=FILE_CLASSIFICATION_MESSAGE,
-                                                    input_variables=["qname", "files"]) \
-                                .format(qname=qname, files="\n".join(component.assigned_files))
-                            file_assignment = self._parse_invoke(prompt, FilePath)
-                            logger.info(
-                                f"[Reference Resolution] LLM matched {reference.qualified_name} in {lang} at {file_assignment.file_path}")
-                            reference.reference_file = file_assignment.file_path
-                            reference.reference_start_line = file_assignment.start_line
-                            reference.reference_end_line = file_assignment.end_line
-                            if reference.reference_file is not None:
-                                matched = True
-                                break
-                if not matched:
-                    logger.error(
-                        f"[Reference Resolution] Reference file could not be resolved for {reference.qualified_name} in any language.")
-        return analysis
 
     def _try_parse(self, message_content, parser):
         try:
