+++ agents/details_agent.py
@@ -86,8 +86,7 @@ class DetailsAgent(CodeBoardingAgent):
             meta_context=meta_context_str,
             project_type=project_type
         )
-        analysis = self._parse_invoke(prompt, AnalysisInsights)
-        return self.fix_source_code_reference_lines(analysis)
+        return self._parse_invoke(prompt, AnalysisInsights)
 
     def apply_feedback(self, analysis: AnalysisInsights, feedback: ValidationInsights):
         """
@@ -99,17 +98,17 @@ class DetailsAgent(CodeBoardingAgent):
         analysis = self._parse_invoke(prompt, AnalysisInsights)
         return self.fix_source_code_reference_lines(analysis)
 
-    def run(self, cfg_str: str, component: Component):
+    def run(self, component: Component):
         """
         Run the details analysis for the given component.
         This method should execute the steps in order and return the final analysis.
         """
-        self.step_subcfg(cfg_str, component)
+        logger.info(f"Processing component: {component.name}")
+        self.step_subcfg(component)
         self.step_cfg(component)
         self.step_enhance_structure(component)
         analysis = self.step_analysis(component)
-
-        return analysis
+        return self.fix_source_code_reference_lines(analysis)
 
     def classify_files(self, component: Component, analysis: AnalysisInsights):
         """
